ci: support challenges at arbitrary nested levels (#32)
With this PR, a challenge is now simply a directory that contains a
directory named "challenge".
In doing so, we can now support that challenge being at an arbitrary
nested level.
For example we might have challenges at
`challenges/legacy/intro-to-cybersecurity/access-control/level-1` or
just `challenges/access-control-level-1`.
This allows for better challenge namespacing.