ci: refactor list-changed-challenges (#34)
The prior implementation is super slow. This PR uses our current
definition that a challenge is simply a directory which immediately
contains a directory named "challenge". If any files in a challenge
directory (for example, including tests) change, the challenge has
changed. We also walk all ancestors back up to the root challenges
directory, looking for "common" directories. If any files inside one of
these "common" directory changes, the challenge has changed.

This significantly improves performance, but will over-estimate on the
"common" directories case (we no longer try to more intelligently
analyze if the common-dependency is actually used).