pwnshop: fix challenge detection for nested challenge directories (#55)
We have defined a challenge to be any directory that contains a
directory named `challenge/`. However,
`challenges/legacy/welcome/welcome/` contains a directory named
`challenge/`. But `welcome` is not actually a challenge, it is a module
named `welcome` which contains a challenge named `challenge`. In other
words the directory
`challenges/legacy/welcome/welcome/challenge/challenge` exists (notice
the double nested `challenge/` directories).

This PR changes the definition so that a challenge (directory) cannot
contain other challenges (directories). Only the deepest node in the
file system is *actually* a challenge.